Move jobs, reduce logging, improve gas station test
